requirements.txt
ollama~=0.4.8
qdrant-client~=1.14.2
fastembed~=0.7.0
fastapi[standard]~=0.115.12
mariadb~=1.1.12
redis~=6.1.1
pymongo~=4.13.2
pyyaml~=6.0.2